refactor: move search into indexing domain
