+++ src/kodit/indexing/indexing_repository.py
@@ -196,3 +196,23 @@ class IndexRepository:
         """
         self.session.add(embedding)
         await self.session.commit()
+
+    async def list_snippets_by_ids(self, ids: list[int]) -> list[tuple[File, Snippet]]:
+        """List snippets by IDs.
+
+        Returns:
+            A list of snippets in the same order as the input IDs.
+
+        """
+        query = (
+            select(Snippet, File)
+            .where(Snippet.id.in_(ids))
+            .join(File, Snippet.file_id == File.id)
+        )
+        rows = await self.session.execute(query)
+
+        # Create a dictionary for O(1) lookup of results by ID
+        id_to_result = {snippet.id: (file, snippet) for snippet, file in rows.all()}
+
+        # Return results in the same order as input IDs
+        return [id_to_result[i] for i in ids]
