+++ src/kodit/indexing/indexing_service.py
@@ -13,11 +13,16 @@ import pydantic
 import structlog
 from tqdm.asyncio import tqdm
 
-from kodit.bm25.keyword_search_service import BM25Document, KeywordSearchProvider
+from kodit.bm25.keyword_search_service import (
+    BM25Document,
+    BM25Result,
+    KeywordSearchProvider,
+)
 from kodit.embedding.vector_search_service import (
     VectorSearchRequest,
     VectorSearchService,
 )
+from kodit.indexing.fusion import reciprocal_rank_fusion
 from kodit.indexing.indexing_models import Snippet
 from kodit.indexing.indexing_repository import IndexRepository
 from kodit.snippets.snippets import SnippetService
@@ -40,6 +45,26 @@ class IndexView(pydantic.BaseModel):
     updated_at: datetime | None = None
     source: str | None = None
     num_snippets: int | None = None
+
+
+class SearchRequest(pydantic.BaseModel):
+    """Request for a search."""
+
+    code_query: str | None = None
+    keywords: list[str] | None = None
+    top_k: int = 10
+
+
+class SearchResult(pydantic.BaseModel):
+    """Data transfer object for search results.
+
+    This model represents a single search result, containing both the file path
+    and the matching snippet content.
+    """
+
+    id: int
+    uri: str
+    content: str
 
 
 class IndexService:
@@ -154,6 +179,58 @@ class IndexService:
 
         # Update index timestamp
         await self.repository.update_index_timestamp(index)
+
+    async def search(self, request: SearchRequest) -> list[SearchResult]:
+        """Search for relevant data."""
+        fusion_list = []
+        if request.keywords:
+            # Gather results for each keyword
+            result_ids: list[BM25Result] = []
+            for keyword in request.keywords:
+                results = await self.keyword_search_provider.retrieve(
+                    keyword, request.top_k
+                )
+                result_ids.extend(results)
+
+            # Sort results by score
+            result_ids.sort(key=lambda x: x[1], reverse=True)
+
+            self.log.debug("Search results (BM25)", results=result_ids)
+
+            bm25_results = [x[0] for x in result_ids]
+            fusion_list.append(bm25_results)
+
+        # Compute embedding for semantic query
+        semantic_results = []
+        if request.code_query:
+            query_embedding = await self.code_search_service.retrieve(
+                request.code_query, top_k=request.top_k
+            )
+            semantic_results = [x.snippet_id for x in query_embedding]
+            fusion_list.append(semantic_results)
+
+        if len(fusion_list) == 0:
+            return []
+
+        # Combine all results together with RFF if required
+        final_results = reciprocal_rank_fusion(fusion_list, k=60)
+
+        # Extract ids from final results
+        final_ids = [x[0] for x in final_results]
+
+        # Get snippets from database (up to top_k)
+        search_results = await self.repository.list_snippets_by_ids(
+            final_ids[: request.top_k]
+        )
+
+        return [
+            SearchResult(
+                id=snippet.id,
+                uri=file.uri,
+                content=snippet.content,
+            )
+            for file, snippet in search_results
+        ]
 
     async def _create_snippets(
         self,
