--- src/kodit/search/__init__.py
@@ -1,1 +1,0 @@
-"""Search for relevant snippets."""
