--- tests/kodit/search/__init__.py
@@ -1,1 +1,0 @@
-"""Test the search module."""
